requirements.txt
requests>=2.25.0
python-dotenv>=1.0.0
pyyaml>=6.0
